+++ config/config.exs
@@ -1,3 +1,6 @@
 import Config
 
-config :dotfiles, dots_path: "/dots"
+config :dotfiles, :paths,
+  dots_path: Path.expand("./dots"),
+  config_path: Path.expand("~/.config"),
+  home_path: Path.expand("~")
